글 삭제시 권한체크 코드 추가

+++ src/main/java/com/back/domain/post/post/controller/ApiV1PostController.java
@@ -52,8 +52,17 @@ public class ApiV1PostController {
     @DeleteMapping("/{id}")
     @Transactional
     @Operation(summary = "삭제")
-    public RsData<Void> delete(@PathVariable int id) {
+    public RsData<Void> delete(
+            @PathVariable int id,
+            @NotBlank @Size(min = 30, max = 50) @RequestHeader("Authorization") String authorization
+    ) {
+        String apiKey = authorization.replace("Bearer ", "");
+        Member actor = memberService.findByApiKey(apiKey).orElseThrow(() -> new ServiceException("401-1", "존재하지 않는 apiKey 입니다."));
+
         Post post = postService.findById(id).get();
+
+        if (!actor.equals(post.getAuthor()))
+            throw new ServiceException("403-1", "글 삭제 권한이 없습니다.");
 
         postService.delete(post);
 
@@ -111,7 +120,6 @@ public class ApiV1PostController {
             @Valid @RequestBody PostModifyReqBody reqBody,
             @NotBlank @Size(min = 30, max = 50) @RequestHeader("Authorization") String authorization
     ) {
-
         String apiKey = authorization.replace("Bearer ", "");
         Member actor = memberService.findByApiKey(apiKey)
                 .orElseThrow(() -> new ServiceException("401-1", "존재하지 않는 apiKey 입니다."));
